Add files via upload
The game is basically trying to get the basketball to touch the basket. Every time it touches the backboard, the player gets one point. However, After 5 few seconds of getting the first point, 5 bananas appear and make getting the ball to the basket a bit more difficult. And every time the banana touches the Cat sprite, the sprite relocates away from the ball. Player uses the arrow keys to move around.